core java

+++ AreaOfTriangle.java
@@ -1,5 +1,5 @@
 import java.util.Scanner;
-import java.math.*;
+
 
 public class AreaOfTriangle {
     public static void main(String[] args) {
@@ -9,7 +9,7 @@ public class AreaOfTriangle {
         float b=sc.nextFloat();
         float c=sc.nextFloat();
         float s=(float)1/2*(a+b+c);
-        float area=(float)Math.sqrt(s*(s-a)*(s-b)*(s-c));
+        double area=Math.sqrt(s*(s-a)*(s-b)*(s-c));  // sqrt() returns double type value.
         System.out.println("Area of the triangle given:"+area);
